ddr_generator/requirements.txt
# DDR Generation System - Python Dependencies

# ========== Core PDF Processing ==========
PyMuPDF==1.23.9
pdfplumber==0.10.3
pillow==10.1.0

# ========== Data Validation ==========
pydantic==2.5.2
pydantic-settings==2.1.0

# ========== LLM Integration ==========
openai==1.6.1
google-generativeai==0.3.2
anthropic==0.8.1

# ========== NLP & Embeddings ==========
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2

# ========== API (Optional) ==========
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# ========== Utilities ==========
python-dotenv==1.0.0
tqdm==4.66.1
rich==13.7.0

# ========== Testing ==========
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# ========== Development ==========
black==23.12.1
flake8==7.0.0
mypy==1.7.1
